refactor to prepare for flag
